// components/common/services_block/enums.ts
export enum CardType {
  PRIMARY,
  SECONDARY,
}

export enum PrimaryContent {
  HEADER = "Чек-листы и\u00A0документы для аренды",
  DESC = "Договора защищающие интересы арендатора. Чек-листы по\u00A0выбору объекта, осмотру и\u00A0переговорам помогут не\u00A0пропустить важное",
  BUTTON_TEXT = "Купить за 799 ₽",
}

export enum SecondaryContent {
  HEADER = "Пакет бесплатных документов",
  DESC = "Пригодятся от\u00A0подписания договора аренды\nдо\u00A0его прекращения. Документы универсальны для всех участников сделки",
  BUTTON_TEXT = "Получить документы",
}

export enum MakeDeal {
  HEADER = "Составление договоров",
  BUTTON_TEXT = "Заказать от 1 500 ₽",
}

export enum ExmDeal {
  HEADER = "Типовые договора",
  BUTTON_TEXT = "Получить документы бесплатно",
}